Extract the debugger console component to a common location, and use it from there.
